+++ deepctr/models/pnn.py
@@ -7,14 +7,10 @@ Reference:
     [1] Qu Y, Cai H, Ren K, et al. Product-based neural networks for user response prediction[C]//Data Mining (ICDM), 2016 IEEE 16th International Conference on. IEEE, 2016: 1149-1154.(https://arxiv.org/pdf/1611.00144.pdf)
 """
 
-from tensorflow.python.keras.layers import Dense, Embedding, Concatenate, Reshape, Flatten
-from tensorflow.python.keras.models import Model
-from tensorflow.python.keras.initializers import RandomNormal
-from tensorflow.python.keras.regularizers import l2
-
-
+import tensorflow as tf
 from ..layers import PredictionLayer, MLP, InnerProductLayer, OutterProductLayer
-from ..utils import get_input
+from ..input_embedding import get_inputs_embedding
+from ..utils import concat_fun
 
 
 def PNN(feature_dim_dict, embedding_size=8, hidden_size=(128, 128), l2_reg_embedding=1e-5, l2_reg_deep=0,
@@ -43,48 +39,35 @@ def PNN(feature_dim_dict, embedding_size=8, hidden_size=(128, 128), l2_reg_embed
             "feature_dim must be a dict like {'sparse':{'field_1':4,'field_2':3,'field_3':2},'dense':['field_5',]}")
     if kernel_type not in ['mat', 'vec', 'num']:
         raise ValueError("kernel_type must be mat,vec or num")
-    sparse_input, dense_input = get_input(feature_dim_dict, None)
-    sparse_embedding = [Embedding(feature_dim_dict["sparse"][feat], embedding_size,
-                                  embeddings_initializer=RandomNormal(
-        mean=0.0, stddev=init_std, seed=seed),
-        embeddings_regularizer=l2(
-        l2_reg_embedding),
-        name='sparse_emb_' + str(i) + '-' + feat) for i, feat in
-        enumerate(feature_dim_dict["sparse"])]
+    deep_emb_list, _, inputs_list = get_inputs_embedding(
+        feature_dim_dict, embedding_size, l2_reg_embedding, 0, init_std, seed, False)
 
-    embed_list = [sparse_embedding[i](sparse_input[i])
-                  for i in range(len(feature_dim_dict["sparse"]))]
-
-    if len(dense_input) > 0:
-        continuous_embedding_list = list(
-            map(Dense(embedding_size, use_bias=False, kernel_regularizer=l2(l2_reg_embedding), ),
-                dense_input))
-        continuous_embedding_list = list(
-            map(Reshape((1, embedding_size)), continuous_embedding_list))
-        embed_list += continuous_embedding_list
-
-    inner_product = Flatten()(InnerProductLayer()(embed_list))
-    outter_product = OutterProductLayer(kernel_type)(embed_list)
+    inner_product = tf.keras.layers.Flatten()(InnerProductLayer()(deep_emb_list))
+    outter_product = OutterProductLayer(kernel_type)(deep_emb_list)
 
     # ipnn deep input
-    linear_signal = Reshape(
-        [len(embed_list)*embedding_size])(Concatenate()(embed_list))
+    linear_signal = tf.keras.layers.Reshape(
+        [len(deep_emb_list)*embedding_size])(concat_fun(deep_emb_list))
 
     if use_inner and use_outter:
-        deep_input = Concatenate()(
+        deep_input = tf.keras.layers.Concatenate()(
             [linear_signal, inner_product, outter_product])
     elif use_inner:
-        deep_input = Concatenate()([linear_signal, inner_product])
+        deep_input = tf.keras.layers.Concatenate()(
+            [linear_signal, inner_product])
     elif use_outter:
-        deep_input = Concatenate()([linear_signal, outter_product])
+        deep_input = tf.keras.layers.Concatenate()(
+            [linear_signal, outter_product])
     else:
         deep_input = linear_signal
 
     deep_out = MLP(hidden_size, activation, l2_reg_deep, keep_prob,
                    False, seed)(deep_input)
-    deep_logit = Dense(1, use_bias=False, activation=None)(deep_out)
-    final_logit = deep_logit
-    output = PredictionLayer(final_activation)(final_logit)
-    model = Model(inputs=sparse_input + dense_input,
-                  outputs=output)
+    deep_logit = tf.keras.layers.Dense(
+        1, use_bias=False, activation=None)(deep_out)
+
+    output = PredictionLayer(final_activation)(deep_logit)
+
+    model = tf.keras.models.Model(inputs=inputs_list,
+                                  outputs=output)
     return model