+++ deepctr/utils.py
@@ -1,10 +1,10 @@
+import collections
 import json
 import logging
 from threading import Thread
 
 import requests
-from tensorflow.python.keras.initializers import RandomNormal
-from tensorflow.python.keras.layers import Embedding, Input
+from tensorflow.python.keras.layers import Dense, Concatenate, add
 
 from .activations import *
 from .layers import *
@@ -31,52 +31,36 @@ custom_objects = {'InnerProductLayer': InnerProductLayer,
                   'InteractingLayer': InteractingLayer}
 
 
-def get_input(feature_dim_dict, bias_feature_dim_dict=None):
-    sparse_input = [Input(shape=(1,), name='sparse_' + str(i) + '-' + feat) for i, feat in
-                    enumerate(feature_dim_dict["sparse"])]
-    dense_input = [Input(shape=(1,), name='dense_' + str(i) + '-' + feat) for i, feat in
-                   enumerate(feature_dim_dict["dense"])]
-    if bias_feature_dim_dict is None:
-        return sparse_input, dense_input
+VarLenFeature = collections.namedtuple(
+    'VarLenFeatureConfig', ['name', 'dimension', 'maxlen', 'combiner'])
+
+
+def get_linear_logit(linear_term, dense_input_, l2_reg):
+    if len(linear_term) > 1:
+        linear_term = add(linear_term)
+    elif len(linear_term) == 1:
+        linear_term = linear_term[0]
     else:
-        bias_sparse_input = [Input(shape=(1,), name='bias_sparse_' + str(i) + '-' + feat) for i, feat in
-                             enumerate(bias_feature_dim_dict["sparse"])]
-        bias_dense_input = [Input(shape=(1,), name='bias_dense_' + str(i) + '-' + feat) for i, feat in
-                            enumerate(bias_feature_dim_dict["dense"])]
-        return sparse_input, dense_input, bias_sparse_input, bias_dense_input
+        linear_term = None
 
+    dense_input = list(dense_input_.values())
+    if len(dense_input) > 0:
+        dense_input__ = dense_input[0] if len(
+            dense_input) == 1 else Concatenate()(dense_input)
+        linear_dense_logit = Dense(
+            1, activation=None, use_bias=False, kernel_regularizer=l2(l2_reg))(dense_input__)
+        if linear_term is not None:
+            linear_term = add([linear_dense_logit, linear_term])
+        else:
+            linear_term = linear_dense_logit
 
-def get_share_embeddings(feature_dim_dict, embedding_size, init_std, seed, l2_rev_V, l2_reg_w):
-    sparse_embedding = [Embedding(feature_dim_dict["sparse"][feat], embedding_size,
-                                  embeddings_initializer=RandomNormal(
-                                      mean=0.0, stddev=init_std, seed=seed),
-                                  embeddings_regularizer=l2(l2_rev_V),
-                                  name='sparse_emb_' + str(i) + '-' + feat) for i, feat in
-                        enumerate(feature_dim_dict["sparse"])]
-    linear_embedding = [Embedding(feature_dim_dict["sparse"][feat], 1,
-                                  embeddings_initializer=RandomNormal(mean=0.0, stddev=init_std,
-                                                                      seed=seed), embeddings_regularizer=l2(l2_reg_w),
-                                  name='linear_emb_' + str(i) + '-' + feat) for
-                        i, feat in enumerate(feature_dim_dict["sparse"])]
+    return linear_term
 
-    return sparse_embedding, linear_embedding
-
-
-def get_sep_embeddings(deep_feature_dim_dict, wide_feature_dim_dict, embedding_size, init_std, seed, l2_rev_V, l2_reg_w):
-    sparse_embedding = [Embedding(deep_feature_dim_dict["sparse"][feat], embedding_size,
-                                  embeddings_initializer=RandomNormal(
-                                      mean=0.0, stddev=init_std, seed=seed),
-                                  embeddings_regularizer=l2(l2_rev_V),
-                                  name='sparse_emb_' + str(i) + '-' + feat) for i, feat in
-                        enumerate(deep_feature_dim_dict["sparse"])]
-    linear_embedding = [Embedding(wide_feature_dim_dict["sparse"][feat], 1,
-                                  embeddings_initializer=RandomNormal(mean=0.0, stddev=init_std,
-                                                                      seed=seed), embeddings_regularizer=l2(l2_reg_w),
-                                  name='linear_emb_' + str(i) + '-' + feat) for
-                        i, feat in enumerate(wide_feature_dim_dict["sparse"])]
-
-    return sparse_embedding, linear_embedding
-
+def concat_fun(inputs,axis=-1):
+    if len(inputs) == 1:
+        return inputs[0]
+    else:
+        return Concatenate(axis=axis)(inputs)
 
 def check_version(version):
     """Return version of package on pypi.python.org using json."""