+++ tests/models/DCN_test.py
@@ -1,34 +1,19 @@
-import numpy as np
 import pytest
 from deepctr.models import DCN
-from ..utils import check_model
+from ..utils import check_model, get_test_data
 
 
 @pytest.mark.parametrize(
     'embedding_size,cross_num,hidden_size,sparse_feature_num',
-    [(8, 0, (32,), 2), (8, 1, (), 1), ('auto', 1, (32,), 3)
+    [(8, 0, (32,), 2), ('auto', 1, (), 1), ('auto', 1, (32,), 3)
      ]
 )
 def test_DCN(embedding_size, cross_num, hidden_size, sparse_feature_num):
     model_name = "DCN"
 
     sample_size = 64
-    feature_dim_dict = {"sparse": {}, 'dense': []}
-    for name, num in zip(["sparse", "dense"], [sparse_feature_num, sparse_feature_num]):
-        if name == "sparse":
-            for i in range(num):
-                feature_dim_dict[name][name + '_' +
-                                       str(i)] = np.random.randint(1, 10)
-        else:
-            for i in range(num):
-                feature_dim_dict[name].append(name + '_' + str(i))
-    sparse_input = [np.random.randint(0, dim, sample_size)
-                    for dim in feature_dim_dict['sparse'].values()]
-    dense_input = [np.random.random(sample_size)
-                   for name in feature_dim_dict['dense']]
-
-    y = np.random.randint(0, 2, sample_size)
-    x = sparse_input + dense_input
+    x, y, feature_dim_dict = get_test_data(
+        sample_size, sparse_feature_num, sparse_feature_num)
 
     model = DCN(feature_dim_dict, embedding_size=embedding_size, cross_num=cross_num,
                 hidden_size=hidden_size, keep_prob=0.5, )